+++ requirements.txt
@@ -1,0 +1,2 @@
+requests~=2.25.1
+git-repo~=1.10.3
